+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Added
 - [x] Added **Shipping Rates API**.
 - [x] Added `TransactionService::readyToPay()` method.
 - [x] Added new property to `Address`: `dob`
-- [x] Added new property to `Invoice`: `discounts`, `shipping`, `tax`
+- [x] Added new properties to `Invoice`: `discounts`, `shipping`, `tax`
 - [x] Added new properties to `Subscription`: `currency`, `voidTime`, `shipping`
 
 ### Deprecated